Update torrc
This update should make things a lot better. The ExtORPort allows the Tor Project to collect safe stats on how your bridge is used. Comment that line out if you don't want to do that. The other tweaks are to the ORPort, for which a random high port is recommended. Change both it and ExtORPort in this file and in the iptables rules files before running the bootstrap script. ServerTransportListenAddr is set to 443 because that's the port the user's traffic will actually go through.